03-chatbot-local-streamlit-ollama/requirements.txt
# Dependencias principales
streamlit==1.28.0
requests==2.31.0

# Para RAG (Retrieval Augmented Generation)
chromadb==0.4.22
sentence-transformers==2.2.2

# Procesamiento de documentos
PyPDF2==3.0.1
python-docx==1.1.0

# Opcional: Ollama Python client
ollama==0.1.7

# Para desarrollo
python-dotenv==1.0.1
